third_party/WebKit/LayoutTests/css3/supports-dom-api-expected.txt

Test window.CSS.supports()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS CSS.supports("display: none") is true
PASS CSS.supports("  display: none ") is true
PASS CSS.supports("(display: none)") is true
PASS CSS.supports("(display: deadbeef)") is false
PASS CSS.supports("display: deadbeef") is false
PASS CSS.supports("(display: none) and ((display: block) or (display: inline))") is true
PASS CSS.supports("(not (display: deadbeef)) and (display: block)") is true
PASS CSS.supports("top: calc(80% - 20px)") is true
PASS CSS.supports("background-color: rgb(0, 128, 0)") is true
PASS CSS.supports("background: url('/blah')") is true
PASS CSS.supports("background: invalid('/blah')") is false
PASS CSS.supports("display: none;") is false
PASS CSS.supports("display: none; garbage") is false
PASS CSS.supports("  display: none ; garbage ") is false
PASS CSS.supports("not (display: deadbeef)") is true
PASS CSS.supports("not (display: none)") is false
PASS CSS.supports("not (not (display: none))") is true
PASS CSS.supports("not (not (not (display: none)))") is false
PASS CSS.supports("(display: none) and (display: block)") is true
PASS CSS.supports("(display: none) and (display: block) and (display: inline)") is true
PASS CSS.supports("(display: none) and (display: block) and (display: deadbeef) and (display: inline)") is false
PASS CSS.supports("(display: none) or (display: inline)") is true
PASS CSS.supports("(display: none) or (display: block) or (display: inline)") is true
PASS CSS.supports("(display: none) or (display: deadbeef) or (display: inline)") is true
PASS CSS.supports("(display: ohhai) or (display: deadbeef) or (display: rainbows)") is false
PASS CSS.supports("(display: none) and (display: block) or (display: inline)") is false
PASS CSS.supports("not (display: deadbeef) and (display: block)") is false
PASS CSS.supports("(not (border: 1px 1px 1px 1px 1px solid #000)) and (display: block)") is true
PASS CSS.supports("(display: block !important) and ((display: inline) or (display: deadbeef))") is true
PASS CSS.supports("not ((not (display: block)) or ((display: none) and (deadbeef: 1px)))") is true
PASS CSS.supports("not( display: deadbeef)") is false
PASS CSS.supports("(display: none)and (   -webkit-transition: all 1s )") is false
PASS CSS.supports("(display: none)or(-webkit-transition: all 1s)") is false
PASS CSS.supports("(display: none) or(-webkit-transition: all 1s    )") is false
PASS CSS.supports("(((((((display: none)))))))") is true
PASS CSS.supports("(!important)") is false
PASS CSS.supports("not not not not (display: none)") is false
PASS CSS.supports("(top: -webkit-calc(80% - 20px))") is true
PASS CSS.supports("(background-color: rgb(0, 128, 0))") is true
PASS CSS.supports("(background: url('/blah'))") is true
PASS CSS.supports("((top: -webkit-calc(80% - 20px)) and (not (background-color: rgb(0, 128, 0)))) or (background: url('/blah'))") is true
PASS CSS.supports("(background: invalid('/blah'))") is false
PASS CSS.supports("top", "20%") is true
PASS CSS.supports("top", " 20px   ") is true
PASS CSS.supports("top", "20") is false
PASS CSS.supports("dis\nplay", "none") is false
PASS CSS.supports("display", "rainbow") is false
PASS CSS.supports("display", "url(rainbow.png)") is false
PASS CSS.supports(" background   ", "url(rainbow.png)") is false
PASS CSS.supports("background", "invalid(rainbow.png)") is false
PASS CSS.supports("background-color", "#000") is true
PASS CSS.supports("background-color", "rgba(0,0,0,0)") is true
PASS CSS.supports("background-color", "inherit") is true
PASS CSS.supports("font-family", "'\n'") is false
PASS CSS.supports("display", "none !important") is false
PASS CSS.supports("display", "none!important") is false
PASS CSS.supports("display", "!important none") is false
PASS CSS.supports("display", "none !important ! important ") is false
PASS CSS.supports("display", "none ! important") is false
PASS CSS.supports("display", "none ! 	important") is false
PASS CSS.supports("display", "none ! \nimportant") is false
PASS CSS.supports("", "") is false
PASS CSS.supports(true, "") is false
PASS CSS.supports([], "none") is false
PASS CSS.supports("display", "") is false
PASS CSS.supports("display:", "none") is false
PASS CSS.supports("z-index", 1) is true
PASS CSS.supports("content", []) is false
PASS CSS.supports("content", "!important") is false
PASS CSS.supports("(display: none)", undefined) is false
PASS CSS.supports("margin", "1__qem") is false
PASS CSS.supports("border", "1px solid #000") is true
PASS CSS.supports("border", "inherit") is true
PASS CSS.supports("border", "1px solid inherit") is false
PASS CSS.supports("border", "inherit solid #000") is false
PASS CSS.supports("border-bottom", "thick green") is true
PASS CSS.supports("border-bottom", "inherit") is true
PASS CSS.supports("border-bottom", "thick green inherit") is false
PASS CSS.supports("border-bottom", "inherit thick green") is false
PASS CSS.supports("border-left", "thick green") is true
PASS CSS.supports("border-left", "inherit") is true
PASS CSS.supports("border-left", "thick green inherit") is false
PASS CSS.supports("border-left", "inherit thick green") is false
PASS CSS.supports("border-right", "thick green") is true
PASS CSS.supports("border-right", "inherit") is true
PASS CSS.supports("border-right", "thick green inherit") is false
PASS CSS.supports("border-right", "inherit thick green") is false
PASS CSS.supports("border-top", "thick green") is true
PASS CSS.supports("border-top", "inherit") is true
PASS CSS.supports("border-top", "thick green inherit") is false
PASS CSS.supports("border-top", "inherit thick green") is false
PASS CSS.supports("border-radius", "1px 0 3px 4px") is true
PASS CSS.supports("border-radius", "inherit") is true
PASS CSS.supports("border-radius", "1px 0 3px inherit") is false
PASS CSS.supports("border-radius", "inherit 0 3px 4px") is false
PASS CSS.supports("border-spacing", "5px 5px") is true
PASS CSS.supports("border-spacing", "inherit") is true
PASS CSS.supports("border-spacing", "5px inherit") is false
PASS CSS.supports("border-spacing", "inherit 5px") is false
PASS CSS.supports("font", "italic small-caps bolder 16px/3 cursive") is true
PASS CSS.supports("font", "inherit") is true
PASS CSS.supports("font", "italic small-caps bolder 16px/3 inherit") is false
PASS CSS.supports("font", "inherit small-caps bolder 16px/3 cursive") is false
PASS CSS.supports("list-style", "georgian inside") is true
PASS CSS.supports("list-style", "inherit") is true
PASS CSS.supports("list-style", "georgian inherit") is false
PASS CSS.supports("list-style", "inherit inside") is false
PASS CSS.supports("margin", "5px") is true
PASS CSS.supports("margin", "inherit") is true
PASS CSS.supports("margin", "inherit 5px") is false
PASS CSS.supports("margin", "5px inherit") is false
PASS CSS.supports("margin", "inherit 5px 5px") is false
PASS CSS.supports("margin", "inherit 5px 5px 5px") is false
PASS CSS.supports("outline", "1px solid #000") is true
PASS CSS.supports("outline", "inherit") is true
PASS CSS.supports("outline", "1px solid inherit") is false
PASS CSS.supports("outline", "inherit solid #000") is false
PASS CSS.supports("overflow", "scroll") is true
PASS CSS.supports("overflow", "inherit") is true
PASS CSS.supports("overflow", "inherit scroll") is false
PASS CSS.supports("overflow", "scroll inherit") is false
PASS CSS.supports("transform", "scaleX(2)") is true
PASS CSS.supports("transform", "inherit") is true
PASS CSS.supports("transform", "scaleX(2) inherit") is false
PASS CSS.supports("transform", "inherit scaleX(2)") is false
PASS CSS.supports("transition", "margin-left 4s") is true
PASS CSS.supports("transition", "inherit") is true
PASS CSS.supports("transition", "margin-left inherit") is false
PASS CSS.supports("transition", "inherit 4s") is false
PASS successfullyParsed is true

TEST COMPLETE
